tests/submissions/anssik/battery-chargingchange.html

<!DOCTYPE html>
<html>
  <head>
  <title>Battery Status API Test Suite</title>
  <script src="http://www.w3c-test.org/resources/testharness.js"></script>
  <link rel='stylesheet' href='http://w3c-test.org/resources/testharness.css' media='all'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
       The battery charging state is updated.
    </p>
    <h2>Preconditions</h2>
    <ol>
      <li>
        The device is unplugged from the charger before this test case is run.
      </li>
      <li>
        The battery must not be full or reach full capacity during the time the test is run.
      </li>
    </ol>
    <div id="log"></div>
    <script>
    (function() {

      // this is the pattern used for testing event handlers
      
      /*
      var onclick_test = async_test('onclick on body');
      document.querySelector('body').onclick = onclick_test.step_func(function (e) {
        // assert_xxxx();
        onclick_test.done();
      });
      
      alert('click on body!');
      */

      // the tests below must be validated on a real implementation, to set proper timeouts etc.
      
      var onchargingchange_test = async_test('When the battery charging state is updated, must fire a chargingchange event.');
      navigator.battery.onchargingchange = onchargingchange_test.step_func(function (e) {
        assert_true(navigator.battery.charging, 'The charging attribute must be set to true, if the battery is charging.')
        onchargingchange_test.done();
      });
      
      var onchargingtimechange_test = async_test('When the battery charging time is updated, must fire a chargingtimechange event.');
      navigator.battery.onchargingtimechange = onchargingtimechange_test.step_func(function (e) {
        assert_false(navigator.battery.chargingTime !== Infinity);
        onchargingchange_test.done();
      });
      
      alert('Plug in the charger, and click OK.');

      })();
    </script>
  </body>
</html>
